(PC-23190)[API] chore: extract post price categories test
